refactor: discount, product model 분리
